Fix merge issues

+++ src/main/java/com/lankatrails/lankatrails_backend/dtos/TripBudgetCategoryDto.java
@@ -4,12 +4,14 @@ import com.lankatrails.lankatrails_backend.model.enums.BudgetCategory;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class TripBudgetCategoryDto {
     private Long limitId;
     private BudgetCategory budgetCategory;
-    private Double limitAmount;
-    private Double spentAmount = 0.0;
+    private BigDecimal limitAmount;
+    private BigDecimal spentAmount = BigDecimal.valueOf(0.0);
     private Long tripId;
 }
